+++ v2/internal/interface/executor/generation.go
@@ -2,23 +2,21 @@ package executor
 
 import (
 	"context"
+
+	"fmt"
 	"path/filepath"
 	"sync"
 
-	"github.com/swipe-io/swipe/v2/internal/types"
-
-	"golang.org/x/tools/go/packages"
-
-	"github.com/swipe-io/swipe/v2/internal/usecase/generator"
-
-	"github.com/swipe-io/swipe/v2/internal/usecase/processor"
-
+	"github.com/swipe-io/swipe/v2/internal/errors"
 	"github.com/swipe-io/swipe/v2/internal/importer"
 	"github.com/swipe-io/swipe/v2/internal/option"
+	"github.com/swipe-io/swipe/v2/internal/types"
 	"github.com/swipe-io/swipe/v2/internal/usecase/executor"
 	"github.com/swipe-io/swipe/v2/internal/usecase/factory"
 	"github.com/swipe-io/swipe/v2/internal/usecase/frame"
-	"github.com/swipe-io/swipe/v2/internal/usecase/registry"
+	"github.com/swipe-io/swipe/v2/internal/usecase/generator"
+	"github.com/swipe-io/swipe/v2/internal/usecase/processor"
+	"golang.org/x/tools/go/packages"
 )
 
 type importerer interface {
@@ -26,10 +24,10 @@ type importerer interface {
 }
 
 type generationExecutor struct {
-	r  registry.ProcessorRegistry
-	i  factory.ImporterFactory
-	ff frame.Factory
-	l  *option.Loader
+	processorFactory processor.Factory
+	importerFactory  factory.ImporterFactory
+	frameFactory     frame.Factory
+	optionLoader     *option.Loader
 }
 
 func (e *generationExecutor) processGenerate(pkg *packages.Package, generators []generator.Generator) <-chan executor.GenerateResult {
@@ -37,10 +35,8 @@ func (e *generationExecutor) processGenerate(pkg *packages.Package, generators [
 
 	go func() {
 		var wg sync.WaitGroup
-
 		for _, g := range generators {
 			wg.Add(1)
-
 			go func(g generator.Generator) {
 				defer wg.Done()
 
@@ -68,7 +64,7 @@ func (e *generationExecutor) processGenerate(pkg *packages.Package, generators [
 				generated.PkgPath = pkg.PkgPath
 				generated.OutputPath = filepath.Join(outputDir, g.Filename())
 
-				newImporter := e.i.NewImporter(generated.OutputPath, pkg)
+				newImporter := e.importerFactory.NewImporter(generated.OutputPath, pkg)
 				if g, ok := g.(importerer); ok {
 					g.SetImporter(newImporter)
 				}
@@ -77,7 +73,7 @@ func (e *generationExecutor) processGenerate(pkg *packages.Package, generators [
 					generated.Errs = append(generated.Errs, err)
 					return
 				}
-				fr := e.ff.NewFrame(generated.OutputPath, newImporter, pkg)
+				fr := e.frameFactory.NewFrame(generated.OutputPath, newImporter, pkg)
 				content, err := fr.Frame(g.Bytes())
 				if err != nil {
 					generated.Content = g.Bytes()
@@ -94,43 +90,45 @@ func (e *generationExecutor) processGenerate(pkg *packages.Package, generators [
 }
 
 func (e *generationExecutor) Execute() (results []executor.GenerateResult, errs []error) {
-	opr, errs := e.l.Load()
+	opr, errs := e.optionLoader.Load()
 	if len(errs) > 0 {
 		return nil, errs
 	}
-	var processors []processor.Processor
+	if len(opr.Options) == 0 {
+		return nil, []error{fmt.Errorf("swipe options not found")}
+	}
+	var wg sync.WaitGroup
 	for _, o := range opr.Options {
-		p, err := e.r.NewProcessor(o, opr.ExternalOptions, opr.Data)
+		fn, ok := e.processorFactory.Get(o.Option.Name)
+		if !ok {
+			errs = append(errs, errors.NotePosition(o.Option.Position,
+				fmt.Errorf("unknown option name %s", o.Option.Name)))
+			continue
+		}
+		p, err := fn(o, opr)
 		if err != nil {
 			errs = append(errs, err)
 			continue
 		}
-		processors = append(processors, p)
-	}
-	if len(errs) > 0 {
-		return nil, errs
-	}
-	var wg sync.WaitGroup
-	for _, p := range processors {
 		wg.Add(1)
-		go func(p processor.Processor) {
+		go func(p processor.Processor, pkg *packages.Package) {
 			defer wg.Done()
-			outCh := e.processGenerate(p.Pkg(), p.Generators())
+			outCh := e.processGenerate(pkg, p.Generators(pkg, opr.Data.WorkDir))
 			for generateResult := range outCh {
 				results = append(results, generateResult)
 			}
-		}(p)
+		}(p, o.Pkg)
 	}
 	wg.Wait()
 	return
 }
 
 func NewGenerationExecutor(
-	r registry.ProcessorRegistry,
-	i factory.ImporterFactory,
-	ff frame.Factory,
-	l *option.Loader,
+	processorFactory processor.Factory,
+	importerFactory factory.ImporterFactory,
+	frameFactory frame.Factory,
+	optionLoader *option.Loader,
 ) executor.GenerationExecutor {
-	return &generationExecutor{r: r, i: i, ff: ff, l: l}
+	return &generationExecutor{processorFactory: processorFactory, importerFactory: importerFactory, frameFactory: frameFactory, optionLoader: optionLoader}
 
 }