+++ v2/internal/interface/generator/jsonrpc_server.go
@@ -2,7 +2,6 @@ package generator
 
 import (
 	"context"
-	stdtypes "go/types"
 	"strconv"
 	stdstrings "strings"
 
@@ -136,14 +135,14 @@ func (g *jsonRPCServer) Process(ctx context.Context) error {
 
 	for i := 0; i < g.options.Interfaces().Len(); i++ {
 		iface := g.options.Interfaces().At(i)
-		typeStr := stdtypes.TypeString(iface.Type(), g.i.QualifyPkg)
+		typeStr := iface.LcNameWithPrefix() + "Interface"
 		if i > 0 {
 			g.W(",")
 		}
 
 		if iface.External() {
 			external = true
-			g.W("%s %sOption", iface.LcName(), iface.UcName())
+			g.W("%s %sOption", iface.LcNameWithPrefix(), iface.UcNameWithPrefix())
 		} else {
 			g.W("svc%s %s", iface.UcName(), typeStr)
 		}
@@ -183,7 +182,7 @@ func (g *jsonRPCServer) Process(ctx context.Context) error {
 			}
 
 			for _, m := range iface.Methods() {
-				optName := iface.LcName()
+				optName := iface.LcNameWithPrefix()
 				epFactoryName := "endpointFactory"
 				kitEndpointPkg := g.i.Import("endpoint", "github.com/go-kit/kit/endpoint")
 				transportExtPkg := g.i.Import(iface.ExternalSwipePkg().Name, iface.ExternalSwipePkg().PkgPath)
@@ -230,7 +229,7 @@ func (g *jsonRPCServer) Process(ctx context.Context) error {
 				)
 				g.W(
 					"%[3]s.%[2]sEndpoint = middlewareChain(append(opts.genericEndpointMiddleware, opts.%[1]sEndpointMiddleware...))(%[3]s.%[2]sEndpoint)\n",
-					m.IfaceLcName, m.Name, epSetName,
+					iface.LcNameWithPrefix()+m.Name, m.Name, epSetName,
 				)
 				g.W("}\n")
 			}